backend/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
python-docx
requests
reportlab
